Add Namespace to classes and autoload them

+++ includes/bootstrap.inc.php
@@ -1,48 +1,17 @@
 <?php
-/**
- * All times should be UTC unless overridden by a user's time zone
- */
-date_default_timezone_set("UTC");
+require 'Mail.php'; /* PHP Mail functions (PEAR) */
+require 'Mail/mime.php'; /* PHP Mail functions (PEAR) */
 
-/**
- * Load classes
- */
-require 'pageTitles.inc.php';
+define('BASE_PATH', realpath(__DIR__) . '/../');
+define('APP_BASE', realpath(__DIR__ . '/../app'));
 
-require __DIR__ . '/../includes/classes/CDCMastery.class.php';
-require __DIR__ . '/../includes/classes/UserManager.class.php';
-require __DIR__ . '/../includes/classes/AFSCManager.class.php';
-require __DIR__ . '/../includes/classes/AuthenticationManager.class.php';
-require __DIR__ . '/../includes/classes/AnswerManager.class.php';
-require __DIR__ . '/../includes/classes/AssociationManager.class.php';
-require __DIR__ . '/../includes/classes/BaseManager.class.php';
-require __DIR__ . '/../includes/classes/ConfigurationManager.class.php';
-require __DIR__ . '/../includes/classes/ZebraSessions.class.php';
-require __DIR__ . '/../includes/classes/EmailQueueManager.class.php';
-require __DIR__ . '/../includes/classes/SystemLog.class.php';
-require __DIR__ . '/../includes/classes/SystemLogFilter.class.php';
-require __DIR__ . '/../includes/classes/OfficeSymbolManager.class.php';
-require __DIR__ . '/../includes/classes/QuestionManager.class.php';
-require __DIR__ . '/../includes/classes/RoleManager.class.php';
-require __DIR__ . '/../includes/classes/Router.class.php';
-require __DIR__ . '/../includes/classes/TestManager.class.php';
-require __DIR__ . '/../includes/classes/UserStatisticsModule.class.php';
-require __DIR__ . '/../includes/classes/UserPasswordResetManager.class.php';
-require __DIR__ . '/../includes/classes/UserActivationManager.class.php';
-require __DIR__ . '/../includes/classes/UserAuthorizationQueueManager.class.php';
-require __DIR__ . '/../includes/classes/SystemMessageManager.class.php';
-require __DIR__ . '/../includes/classes/TrainingManagerOverview.class.php';
-require __DIR__ . '/../includes/classes/SupervisorOverview.class.php';
-require __DIR__ . '/../includes/classes/SearchModule.class.php';
-require __DIR__ . '/../includes/classes/StatisticsModule.class.php';
-require __DIR__ . '/../includes/classes/FlashCardManager.class.php';
-require __DIR__ . '/../includes/classes/TestGenerator.class.php';
+require BASE_PATH . '/includes/classes/CDCMastery/AutoLoader.class.php';
 
-/**
- * PHP Mail functions (PEAR)
- */
-require 'Mail.php';
-require 'Mail/mime.php';
+$autoLoader = new CDCMastery\PSR4AutoLoader();
+$autoLoader->addNamespace("CDCMastery",__DIR__ . '/classes/CDCMastery');
+$autoLoader->register();
+
+$app = new CDCMastery\Application();
 
 /**
  * Set configuration parameters
@@ -66,9 +35,9 @@ if($db->connect_errno){
 }
 
 /**
- * Load Memcache
+ *
  */
-$memcache = new Memcache();
+$memcache = new Memcache(); /* Load Memcache */
 $memcache->connect($configurationManager->getMemcachedConfiguration('host'),$configurationManager->getMemcachedConfiguration('port'));
 
 define('ENCRYPTION_KEY', $configurationManager->getEncryptionKey());
@@ -76,40 +45,43 @@ define('ENCRYPTION_KEY', $configurationManager->getEncryptionKey());
 /**
  * Instantiate base classes
  */
-$cdcMastery = new CDCMastery();
-$zebraSession = new ZebraSessions($db,"92304j8j8fjsdsn923enkc");
-$systemMessages = new SystemMessageManager();
-$systemLog = new SystemLog($db);
-$emailQueue = new EmailQueueManager($db, 
+$cdcMastery = new CDCMastery\CDCMastery();
+$zebraSession = new CDCMastery\ZebraSessions($db,"92304j8j8fjsdsn923enkc");
+$systemMessages = new CDCMastery\SystemMessageManager();
+$systemLog = new CDCMastery\SystemLog($db);
+$emailQueue = new CDCMastery\EmailQueueManager($db, 
                                     $systemLog, 
                                     $configurationManager->getMailServerConfiguration('host'),
                                     $configurationManager->getMailServerConfiguration('port'),
                                     $configurationManager->getMailServerConfiguration('username'),
                                     $configurationManager->getMailServerConfiguration('password'));
-$roleManager = new RoleManager($db, $systemLog, $emailQueue);
-$baseManager = new BaseManager($db, $systemLog);
-$afscManager = new AFSCManager($db, $systemLog);
-$userManager = new UserManager($db, $systemLog, $emailQueue);
-$officeSymbolManager = new OfficeSymbolManager($db, $systemLog);
-$userStatistics = new UserStatisticsModule($db, $systemLog, $roleManager, $memcache);
-$associationManager = new AssociationManager($db, $systemLog, $userManager, $afscManager, $emailQueue);
+$roleManager = new CDCMastery\RoleManager($db, $systemLog, $emailQueue);
+$baseManager = new CDCMastery\BaseManager($db, $systemLog);
+$afscManager = new CDCMastery\AFSCManager($db, $systemLog);
+$userManager = new CDCMastery\UserManager($db, $systemLog, $emailQueue);
+$officeSymbolManager = new CDCMastery\OfficeSymbolManager($db, $systemLog);
+$userStatistics = new CDCMastery\UserStatisticsModule($db, $systemLog, $roleManager, $memcache);
+$associationManager = new CDCMastery\AssociationManager($db, $systemLog, $userManager, $afscManager, $emailQueue);
 
 if(isset($_SESSION['userUUID']) && !empty($_SESSION['userUUID'])){
-    /**
-     * Something went very wrong, and if the application cannot load the user, let's just destroy the session and start over
-     */
-	if(!$userManager->loadUser($_SESSION['userUUID'])){
+	if(!$userManager->loadUser($_SESSION['userUUID'])){ /* Something went very wrong, and if the application cannot load the user, let's just destroy the session and start over */
         $zebraSession->destroy(session_id());
         $cdcMastery->redirect("/auth/logout");
     }
 
-    /**
-     * This updates the user's latest recorded activity on each page request
-     */
-    $userManager->updateLastActiveTimestamp();
-
-    /**
-     * Ensure user statistics module is fetching stats for this user
-     */
-	$userStatistics->setUserUUID($_SESSION['userUUID']);
+    $userManager->updateLastActiveTimestamp(); /* This updates the user's latest recorded activity on each page request */
+	$userStatistics->setUserUUID($_SESSION['userUUID']); /* Ensure user statistics module is fetching stats for this user */
 }
+
+$router = new CDCMastery\Router($systemLog,$systemMessages);
+
+if($router->showTheme)
+    include BASE_PATH . '/theme/header.inc.php';
+
+include $router->outputPage;
+
+if($router->showTheme)
+    include BASE_PATH . '/theme/footer.inc.php';
+
+$router->__destruct();
+$app->ApplicationShutdown();
